screen-cast-stream-src: Always advertise support for BGRA
Some clients - notably the Gstreamer vapostproc element when using Intel
GPUs - only support BGRA, not BGRx. We already assume that we can
support this format for window screen casts, and even in case of failure
we now have a re-negotiation fallback in place. Thus it's pretty safe to
support it for all screen cast types.

The possible duplication in case of window screen casts doesn't seem to
be a problem for either Pipewire or existing clients like OBS.

Note that the implementation lays the foundation to make it easy to add
more formats in the future.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3617>