cally/text: Fix emission of `text_caret_moved` signal
CallyText was still listening to changes to the deprecated
`ClutterText:position` property to emit the signal, but the
property was removed in commit 3184986897.

Switch to the proper `cursor-position` property to fix the
issue.

Close: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7442
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3621>