backends/native: Work with old libdrm
commit a775241efd8b840e7789077fb2224f0df2361d54 recently introduced
usage of a sort of new API, drmModeCloseFb. Unfortunately, this API
has not yet made it into freedesktop-sdk.

As an interim answer, this commit adds a drmIoctl based reimplementation
of drmModeCloseFb that only gets used if the real implementation isn't
available in libdrm.

Later we can just drop this function without changing any other code.