clutter/input-device: Add a ClutterInputDevice bus-type property
This will be needed for the next version of the tablet protocol, just
VID and PID are not enough to distinguish some devices.

This is only hooked up in the native backend, X does not provide us with
this information at this point.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3724>