tests/test-runner: Add popup commands and popup test
The popup test checks the stacking order, activation, etc, and generally
makes the stacking tests able to use popups as part of their test cases.

Testing of X11 clients is a bit more limiting, e.g. activation resulting
in a sensible stacking order isn't possible.

The way popups are identified across the test runner and client also
differs depending on whether a client is X11 or Wayland. In the X11
case, we can still set a "title", which we'll use for matching, but on
Wayland, we'll have to derive what MetaWindow matches the created popup
by counting the popup depth, i.e. how many parents between the popup and
the toplevel.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4424>